+++ pkg/provider/provider.go
@@ -42,6 +42,7 @@ type KedaProvider struct {
 	externalMetrics  []externalMetric
 	scaleHandler     scaling.ScaleHandler
 	watchedNamespace string
+	ctx              context.Context
 }
 
 type externalMetric struct{}
@@ -50,13 +51,14 @@ var logger logr.Logger
 var metricsServer prommetrics.PrometheusMetricServer
 
 // NewProvider returns an instance of KedaProvider
-func NewProvider(adapterLogger logr.Logger, scaleHandler scaling.ScaleHandler, client client.Client, watchedNamespace string) provider.MetricsProvider {
+func NewProvider(ctx context.Context, adapterLogger logr.Logger, scaleHandler scaling.ScaleHandler, client client.Client, watchedNamespace string) provider.MetricsProvider {
 	provider := &KedaProvider{
 		values:           make(map[provider.CustomMetricInfo]int64),
 		externalMetrics:  make([]externalMetric, 2, 10),
 		client:           client,
 		scaleHandler:     scaleHandler,
 		watchedNamespace: watchedNamespace,
+		ctx:              ctx,
 	}
 	logger = adapterLogger.WithName("provider")
 	logger.Info("starting")
@@ -92,14 +94,18 @@ func (p *KedaProvider) GetExternalMetric(ctx context.Context, namespace string, 
 	}
 
 	scaledObject := &scaledObjects.Items[0]
-	matchingMetrics := []external_metrics.ExternalMetricValue{}
-	scalers, err := p.scaleHandler.GetScalers(ctx, scaledObject)
+	var matchingMetrics []external_metrics.ExternalMetricValue
+	cache, err := p.scaleHandler.GetScalersCache(ctx, scaledObject)
+	if err != nil {
+		return nil, err
+	}
+
 	metricsServer.RecordScalerObjectError(scaledObject.Namespace, scaledObject.Name, err)
 	if err != nil {
 		return nil, fmt.Errorf("error when getting scalers %s", err)
 	}
 
-	for scalerIndex, scaler := range scalers {
+	for scalerIndex, scaler := range cache.GetScalers() {
 		metricSpecs := scaler.GetMetricSpecForScaling(ctx)
 		scalerName := strings.Replace(fmt.Sprintf("%T", scaler), "*scalers.", "", 1)
 
@@ -110,7 +116,8 @@ func (p *KedaProvider) GetExternalMetric(ctx context.Context, namespace string, 
 			}
 			// Filter only the desired metric
 			if strings.EqualFold(metricSpec.External.Metric.Name, info.Metric) {
-				metrics, err := p.getMetricsWithFallback(ctx, scaler, info.Metric, metricSelector, scaledObject, metricSpec)
+				metrics, err := cache.GetMetricsForScaler(ctx, scalerIndex, info.Metric, metricSelector)
+				metrics, err = p.getMetricsWithFallback(ctx, metrics, err, info.Metric, scaledObject, metricSpec)
 
 				if err != nil {
 					logger.Error(err, "error getting metric for scaler", "scaledObject.Namespace", scaledObject.Namespace, "scaledObject.Name", scaledObject.Name, "scaler", scaler)
@@ -124,7 +131,6 @@ func (p *KedaProvider) GetExternalMetric(ctx context.Context, namespace string, 
 				metricsServer.RecordHPAScalerError(namespace, scaledObject.Name, scalerName, scalerIndex, info.Metric, err)
 			}
 		}
-		scaler.Close(ctx)
 	}
 
 	if len(matchingMetrics) == 0 {
@@ -145,7 +151,7 @@ func (p *KedaProvider) ListAllExternalMetrics() []provider.ExternalMetricInfo {
 	opts := []client.ListOption{
 		client.InNamespace(p.watchedNamespace),
 	}
-	err := p.client.List(context.TODO(), scaledObjects, opts...)
+	err := p.client.List(p.ctx, scaledObjects, opts...)
 	if err != nil {
 		logger.Error(err, "Cannot get list of ScaledObjects", "WatchedNamespace", p.watchedNamespace)
 		return nil