+++ pkg/scaling/scale_handler.go
@@ -19,6 +19,7 @@ package scaling
 import (
 	"context"
 	"fmt"
+	"strings"
 	"sync"
 	"time"
 
@@ -34,17 +35,18 @@ import (
 	kedav1alpha1 "github.com/kedacore/keda/v2/apis/keda/v1alpha1"
 	"github.com/kedacore/keda/v2/pkg/eventreason"
 	"github.com/kedacore/keda/v2/pkg/scalers"
+	"github.com/kedacore/keda/v2/pkg/scaling/cache"
 	"github.com/kedacore/keda/v2/pkg/scaling/executor"
 	"github.com/kedacore/keda/v2/pkg/scaling/resolver"
-	"github.com/kedacore/keda/v2/pkg/scaling/scaledjob"
 )
 
 // ScaleHandler encapsulates the logic of calling the right scalers for
 // each ScaledObject and making the final scale decision and operation
 type ScaleHandler interface {
-	HandleScalableObject(scalableObject interface{}) error
-	DeleteScalableObject(scalableObject interface{}) error
-	GetScalers(ctx context.Context, scalableObject interface{}) ([]scalers.Scaler, error)
+	HandleScalableObject(ctx context.Context, scalableObject interface{}) error
+	DeleteScalableObject(ctx context.Context, scalableObject interface{}) error
+	GetScalersCache(ctx context.Context, scalableObject interface{}) (*cache.ScalersCache, error)
+	ClearScalersCache(ctx context.Context, name, namespace string)
 }
 
 type scaleHandler struct {
@@ -54,6 +56,8 @@ type scaleHandler struct {
 	scaleExecutor     executor.ScaleExecutor
 	globalHTTPTimeout time.Duration
 	recorder          record.EventRecorder
+	scalerCaches      map[string]*cache.ScalersCache
+	lock              *sync.RWMutex
 }
 
 // NewScaleHandler creates a ScaleHandler object
@@ -65,24 +69,12 @@ func NewScaleHandler(client client.Client, scaleClient scale.ScalesGetter, recon
 		scaleExecutor:     executor.NewScaleExecutor(client, scaleClient, reconcilerScheme, recorder),
 		globalHTTPTimeout: globalHTTPTimeout,
 		recorder:          recorder,
+		scalerCaches:      map[string]*cache.ScalersCache{},
+		lock:              &sync.RWMutex{},
 	}
 }
 
-func (h *scaleHandler) GetScalers(ctx context.Context, scalableObject interface{}) ([]scalers.Scaler, error) {
-	withTriggers, err := asDuckWithTriggers(scalableObject)
-	if err != nil {
-		return nil, err
-	}
-
-	podTemplateSpec, containerName, err := resolver.ResolveScaleTargetPodSpec(h.client, h.logger, scalableObject)
-	if err != nil {
-		return nil, err
-	}
-
-	return h.buildScalers(ctx, withTriggers, podTemplateSpec, containerName)
-}
-
-func (h *scaleHandler) HandleScalableObject(scalableObject interface{}) error {
+func (h *scaleHandler) HandleScalableObject(ctx context.Context, scalableObject interface{}) error {
 	withTriggers, err := asDuckWithTriggers(scalableObject)
 	if err != nil {
 		h.logger.Error(err, "error duck typing object into withTrigger")
@@ -90,7 +82,7 @@ func (h *scaleHandler) HandleScalableObject(scalableObject interface{}) error {
 	}
 
 	key := withTriggers.GenerateIdenitifier()
-	ctx, cancel := context.WithCancel(context.TODO())
+	ctx, cancel := context.WithCancel(ctx)
 
 	// cancel the outdated ScaleLoop for the same ScaledObject (if exists)
 	value, loaded := h.scaleLoopContexts.LoadOrStore(key, cancel)
@@ -119,7 +111,7 @@ func (h *scaleHandler) HandleScalableObject(scalableObject interface{}) error {
 	return nil
 }
 
-func (h *scaleHandler) DeleteScalableObject(scalableObject interface{}) error {
+func (h *scaleHandler) DeleteScalableObject(ctx context.Context, scalableObject interface{}) error {
 	withTriggers, err := asDuckWithTriggers(scalableObject)
 	if err != nil {
 		h.logger.Error(err, "error duck typing object into withTrigger")
@@ -146,46 +138,89 @@ func (h *scaleHandler) DeleteScalableObject(scalableObject interface{}) error {
 func (h *scaleHandler) startScaleLoop(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, scalableObject interface{}, scalingMutex sync.Locker) {
 	logger := h.logger.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
 
-	// kick off one check to the scalers now
-	h.checkScalers(ctx, scalableObject, scalingMutex)
-
 	pollingInterval := withTriggers.GetPollingInterval()
 	logger.V(1).Info("Watching with pollingInterval", "PollingInterval", pollingInterval)
 
 	for {
 		tmr := time.NewTimer(pollingInterval)
+		h.checkScalers(ctx, scalableObject, scalingMutex)
 
 		select {
 		case <-tmr.C:
-			h.checkScalers(ctx, scalableObject, scalingMutex)
 			tmr.Stop()
 		case <-ctx.Done():
 			logger.V(1).Info("Context canceled")
+			h.ClearScalersCache(ctx, withTriggers.Name, withTriggers.Namespace)
 			tmr.Stop()
 			return
 		}
 	}
 }
 
+func (h *scaleHandler) GetScalersCache(ctx context.Context, scalableObject interface{}) (*cache.ScalersCache, error) {
+	withTriggers, err := asDuckWithTriggers(scalableObject)
+	if err != nil {
+		return nil, err
+	}
+
+	key := strings.ToLower(fmt.Sprintf("%s.%s.%s", withTriggers.Kind, withTriggers.Name, withTriggers.Namespace))
+
+	h.lock.RLock()
+	if cache, ok := h.scalerCaches[key]; ok && cache.Generation == withTriggers.Generation {
+		h.lock.RUnlock()
+		return cache, nil
+	}
+	h.lock.RUnlock()
+
+	h.lock.Lock()
+	defer h.lock.Unlock()
+	if cache, ok := h.scalerCaches[key]; ok && cache.Generation == withTriggers.Generation {
+		return cache, nil
+	} else if ok {
+		cache.Close(ctx)
+	}
+
+	podTemplateSpec, containerName, err := resolver.ResolveScaleTargetPodSpec(ctx, h.client, h.logger, scalableObject)
+	if err != nil {
+		return nil, err
+	}
+
+	scalers := h.buildScalers(ctx, withTriggers, podTemplateSpec, containerName)
+
+	h.scalerCaches[key] = &cache.ScalersCache{
+		Generation: withTriggers.Generation,
+		Scalers:    scalers,
+		Logger:     h.logger,
+		Recorder:   h.recorder,
+	}
+
+	return h.scalerCaches[key], nil
+}
+
+func (h *scaleHandler) ClearScalersCache(ctx context.Context, name, namespace string) {
+	h.lock.Lock()
+	defer h.lock.Unlock()
+
+	key := strings.ToLower(fmt.Sprintf("%s.%s", name, namespace))
+	if cache, ok := h.scalerCaches[key]; ok {
+		cache.Close(ctx)
+		delete(h.scalerCaches, key)
+	}
+}
+
 func (h *scaleHandler) startPushScalers(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, scalableObject interface{}, scalingMutex sync.Locker) {
 	logger := h.logger.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
-	ss, err := h.GetScalers(ctx, scalableObject)
+	cache, err := h.GetScalersCache(ctx, scalableObject)
 	if err != nil {
 		logger.Error(err, "Error getting scalers", "object", scalableObject)
 		return
 	}
 
-	for _, s := range ss {
-		scaler, ok := s.(scalers.PushScaler)
-		if !ok {
-			s.Close(ctx)
-			continue
-		}
-
-		go func() {
+	for _, ps := range cache.GetPushScalers() {
+		go func(s scalers.PushScaler) {
 			activeCh := make(chan bool)
-			go scaler.Run(ctx, activeCh)
-			defer scaler.Close(ctx)
+			go s.Run(ctx, activeCh)
+			defer s.Close(ctx)
 			for {
 				select {
 				case <-ctx.Done():
@@ -201,14 +236,14 @@ func (h *scaleHandler) startPushScalers(ctx context.Context, withTriggers *kedav
 					scalingMutex.Unlock()
 				}
 			}
-		}()
+		}(ps)
 	}
 }
 
 // checkScalers contains the main logic for the ScaleHandler scaling logic.
 // It'll check each trigger active status then call RequestScale
 func (h *scaleHandler) checkScalers(ctx context.Context, scalableObject interface{}, scalingMutex sync.Locker) {
-	scalers, err := h.GetScalers(ctx, scalableObject)
+	cache, err := h.GetScalersCache(ctx, scalableObject)
 	if err != nil {
 		h.logger.Error(err, "Error getting scalers", "object", scalableObject)
 		return
@@ -223,7 +258,7 @@ func (h *scaleHandler) checkScalers(ctx context.Context, scalableObject interfac
 			h.logger.Error(err, "Error getting scaledObject", "object", scalableObject)
 			return
 		}
-		isActive, isError := h.isScaledObjectActive(ctx, scalers, obj)
+		isActive, isError, _ := cache.IsScaledObjectActive(ctx, obj)
 		h.scaleExecutor.RequestScale(ctx, obj, isActive, isError)
 	case *kedav1alpha1.ScaledJob:
 		err = h.client.Get(ctx, types.NamespacedName{Name: obj.Name, Namespace: obj.Namespace}, obj)
@@ -231,83 +266,62 @@ func (h *scaleHandler) checkScalers(ctx context.Context, scalableObject interfac
 			h.logger.Error(err, "Error getting scaledJob", "object", scalableObject)
 			return
 		}
-		isActive, scaleTo, maxScale := h.isScaledJobActive(ctx, scalers, obj)
+		isActive, scaleTo, maxScale := cache.IsScaledJobActive(ctx, obj)
 		h.scaleExecutor.RequestJobScale(ctx, obj, isActive, scaleTo, maxScale)
 	}
 }
 
-func (h *scaleHandler) isScaledObjectActive(ctx context.Context, scalers []scalers.Scaler, scaledObject *kedav1alpha1.ScaledObject) (bool, bool) {
-	isActive := false
-	isError := false
-	for i, scaler := range scalers {
-		isTriggerActive, err := scaler.IsActive(ctx)
-		scaler.Close(ctx)
-
-		if err != nil {
-			h.logger.V(1).Info("Error getting scale decision", "Error", err)
-			isError = true
-			h.recorder.Event(scaledObject, corev1.EventTypeWarning, eventreason.KEDAScalerFailed, err.Error())
-			continue
-		} else if isTriggerActive {
-			isActive = true
-			if externalMetricsSpec := scaler.GetMetricSpecForScaling(ctx)[0].External; externalMetricsSpec != nil {
-				h.logger.V(1).Info("Scaler for scaledObject is active", "Metrics Name", externalMetricsSpec.Metric.Name)
-			}
-			if resourceMetricsSpec := scaler.GetMetricSpecForScaling(ctx)[0].Resource; resourceMetricsSpec != nil {
-				h.logger.V(1).Info("Scaler for scaledObject is active", "Metrics Name", resourceMetricsSpec.Name)
-			}
-			closeScalers(ctx, scalers[i+1:])
-			break
-		}
-	}
-	return isActive, isError
-}
-
-func (h *scaleHandler) isScaledJobActive(ctx context.Context, scalers []scalers.Scaler, scaledJob *kedav1alpha1.ScaledJob) (bool, int64, int64) {
-	return scaledjob.GetScaleMetrics(ctx, scalers, scaledJob, h.recorder)
-}
-
 // buildScalers returns list of Scalers for the specified triggers
-func (h *scaleHandler) buildScalers(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, podTemplateSpec *corev1.PodTemplateSpec, containerName string) ([]scalers.Scaler, error) {
+func (h *scaleHandler) buildScalers(ctx context.Context, withTriggers *kedav1alpha1.WithTriggers, podTemplateSpec *corev1.PodTemplateSpec, containerName string) []cache.ScalerBuilder {
 	logger := h.logger.WithValues("type", withTriggers.Kind, "namespace", withTriggers.Namespace, "name", withTriggers.Name)
-	var scalersRes []scalers.Scaler
 	var err error
 	resolvedEnv := make(map[string]string)
-	if podTemplateSpec != nil {
-		resolvedEnv, err = resolver.ResolveContainerEnv(h.client, logger, &podTemplateSpec.Spec, containerName, withTriggers.Namespace)
-		if err != nil {
-			return scalersRes, fmt.Errorf("error resolving secrets for ScaleTarget: %s", err)
-		}
-	}
+	result := make([]cache.ScalerBuilder, 0, len(withTriggers.Spec.Triggers))
 
-	for scalerIndex, trigger := range withTriggers.Spec.Triggers {
-		config := &scalers.ScalerConfig{
-			Name:              withTriggers.Name,
-			Namespace:         withTriggers.Namespace,
-			TriggerMetadata:   trigger.Metadata,
-			ResolvedEnv:       resolvedEnv,
-			AuthParams:        make(map[string]string),
-			GlobalHTTPTimeout: h.globalHTTPTimeout,
-			ScalerIndex:       scalerIndex,
+	for scalerIndex, t := range withTriggers.Spec.Triggers {
+		triggerName, trigger := scalerIndex, t
+		factory := func() (scalers.Scaler, error) {
+			if podTemplateSpec != nil {
+				resolvedEnv, err = resolver.ResolveContainerEnv(ctx, h.client, logger, &podTemplateSpec.Spec, containerName, withTriggers.Namespace)
+				if err != nil {
+					return nil, fmt.Errorf("error resolving secrets for ScaleTarget: %s", err)
+				}
+			}
+			config := &scalers.ScalerConfig{
+				Name:              withTriggers.Name,
+				Namespace:         withTriggers.Namespace,
+				TriggerMetadata:   trigger.Metadata,
+				ResolvedEnv:       resolvedEnv,
+				AuthParams:        make(map[string]string),
+				GlobalHTTPTimeout: h.globalHTTPTimeout,
+				ScalerIndex:       scalerIndex,
+			}
+
+			config.AuthParams, config.PodIdentity, err = resolver.ResolveAuthRefAndPodIdentity(ctx, h.client, logger, trigger.AuthenticationRef, podTemplateSpec, withTriggers.Namespace)
+			if err != nil {
+				return nil, err
+			}
+
+			return buildScaler(ctx, h.client, trigger.Type, config)
 		}
 
-		config.AuthParams, config.PodIdentity, err = resolver.ResolveAuthRefAndPodIdentity(h.client, logger, trigger.AuthenticationRef, podTemplateSpec, withTriggers.Namespace)
+		scaler, err := factory()
 		if err != nil {
-			closeScalers(ctx, scalersRes)
-			return []scalers.Scaler{}, err
-		}
-
-		scaler, err := buildScaler(ctx, h.client, trigger.Type, config)
-		if err != nil {
-			closeScalers(ctx, scalersRes)
 			h.recorder.Event(withTriggers, corev1.EventTypeWarning, eventreason.KEDAScalerFailed, err.Error())
-			return []scalers.Scaler{}, fmt.Errorf("error getting scaler for trigger #%d: %s", scalerIndex, err)
+			h.logger.Error(err, "error resolving auth params", "scalerIndex", scalerIndex, "object", withTriggers, "trigger", triggerName)
+			if scaler != nil {
+				scaler.Close(ctx)
+			}
+			continue
 		}
 
-		scalersRes = append(scalersRes, scaler)
+		result = append(result, cache.ScalerBuilder{
+			Scaler:  scaler,
+			Factory: factory,
+		})
 	}
 
-	return scalersRes, nil
+	return result
 }
 
 func buildScaler(ctx context.Context, client client.Client, triggerType string, config *scalers.ScalerConfig) (scalers.Scaler, error) {
@@ -334,7 +348,7 @@ func buildScaler(ctx context.Context, client client.Client, triggerType string, 
 	case "azure-queue":
 		return scalers.NewAzureQueueScaler(config)
 	case "azure-servicebus":
-		return scalers.NewAzureServiceBusScaler(config)
+		return scalers.NewAzureServiceBusScaler(ctx, config)
 	case "cassandra":
 		return scalers.NewCassandraScaler(config)
 	case "cpu":
@@ -428,11 +442,5 @@ func asDuckWithTriggers(scalableObject interface{}) (*kedav1alpha1.WithTriggers,
 	default:
 		// here could be the conversion from unknown Duck type potentially in the future
 		return nil, fmt.Errorf("unknown scalable object type %v", scalableObject)
-	}
-}
-
-func closeScalers(ctx context.Context, scalers []scalers.Scaler) {
-	for _, scaler := range scalers {
-		defer scaler.Close(ctx)
 	}
 }